+++ cmd/firewall_update.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/civo/civogo"
 	"github.com/civo/cli/config"
 	"github.com/civo/cli/utility"
 	"github.com/spf13/cobra"
@@ -17,8 +18,17 @@ var firewallUpdateCmd = &cobra.Command{
 	Args:    cobra.MinimumNArgs(2),
 	Run: func(cmd *cobra.Command, args []string) {
 		client, err := config.CivoAPIClient()
+		if regionSet != "" {
+			client.Region = regionSet
+		}
 		if err != nil {
 			utility.Error("Creating the connection to Civo's API failed with %s", err)
+			os.Exit(1)
+		}
+
+		defautlNetwork, err := client.GetDefaultNetwork()
+		if err != nil {
+			utility.Error("%s", err)
 			os.Exit(1)
 		}
 
@@ -28,7 +38,12 @@ var firewallUpdateCmd = &cobra.Command{
 			os.Exit(1)
 		}
 
-		_, err = client.RenameFirewall(firewall.ID, args[1])
+		firewallConfig := &civogo.FirewallConfig{
+			Name:      args[1],
+			NetworkID: defautlNetwork.ID,
+		}
+
+		_, err = client.RenameFirewall(firewall.ID, firewallConfig)
 		if err != nil {
 			utility.Error("%s", err)
 			os.Exit(1)