+++ cmd/instance_create.go
@@ -50,6 +50,9 @@ If you wish to use a custom format, the available fields are:
 	* CreatedAt`,
 	Run: func(cmd *cobra.Command, args []string) {
 		client, err := config.CivoAPIClient()
+		if regionSet != "" {
+			client.Region = regionSet
+		}
 		if err != nil {
 			utility.Error("Creating the connection to Civo's API failed with %s", err)
 			os.Exit(1)