+++ cmd/instance_firewall.go
@@ -29,6 +29,9 @@ If you wish to use a custom format, the available fields are:
 		}
 
 		client, err := config.CivoAPIClient()
+		if regionSet != "" {
+			client.Region = regionSet
+		}
 		if err != nil {
 			utility.Error("Creating the connection to Civo's API failed with %s", err)
 			os.Exit(1)