+++ cmd/network_list.go
@@ -23,6 +23,9 @@ If you wish to use a custom format, the available fields are:
 	* Default`,
 	Run: func(cmd *cobra.Command, args []string) {
 		client, err := config.CivoAPIClient()
+		if regionSet != "" {
+			client.Region = regionSet
+		}
 		if err != nil {
 			utility.Error("Creating the connection to Civo's API failed with %s", err)
 			return
@@ -40,7 +43,7 @@ If you wish to use a custom format, the available fields are:
 			ow.StartLine()
 			ow.AppendData("ID", network.ID)
 			ow.AppendData("Label", network.Label)
-			ow.AppendData("Region", network.Region)
+			ow.AppendData("Region", client.Region)
 			ow.AppendData("CIDR", network.CIDR)
 			ow.AppendData("Default", strconv.FormatBool(network.Default))