+++ cmd/snapshot_list.go
@@ -35,6 +35,9 @@ If you wish to use a custom format, the available fields are:
 Example: civo snapshot ls -o custom -f "ID: Name (Hostname)"`,
 	Run: func(cmd *cobra.Command, args []string) {
 		client, err := config.CivoAPIClient()
+		if regionSet != "" {
+			client.Region = regionSet
+		}
 		if err != nil {
 			utility.Error("Creating the connection to Civo's API failed with %s", err)
 			return